+++ embc-app/ClientApp/src/app/local-authority/pages/local-authority-registrations-page/local-authority-registrations-page.component.html
@@ -1,6 +1,6 @@
 <app-layout-main>
   <!-- TODO: this registration list should be broken into components without the header block -->
-  <app-registration-list></app-registration-list>
+  <app-evacuee-list></app-evacuee-list>
 </app-layout-main>
 
 <app-layout-side>
