+++ embc-app/ClientApp/src/app/provincial-admin/pages/registrations/provincial-admin-registrations-page.component.html
@@ -1,6 +1,6 @@
 <app-layout-main>
   <!-- TODO: this registration list should be broken into components without the header block -->
-  <app-registration-list></app-registration-list>
+  <app-evacuee-list></app-evacuee-list>
 </app-layout-main>
 
 <app-layout-side>
