--- embc-app/ClientApp/src/app/volunteer/components/registration-list/registration-list.component.html
@@ -1,95 +1,0 @@
-<div class="mb-4">
-  <h1>Emergency Support Services Dashboard</h1>
-  <h2>Welcome to your Dashboard!</h2>
-  <p>As an Emergency Support Services user you'll be able to register new evacuees, complete an existing evacuee
-    self-registration, or if needed, search and edit evacuee information.</p>
-  <h3 class="blue-header">LIST OF EVACUEES</h3>
-
-  <div role="search" class="form-row">
-    <div class="col">
-      <input [(ngModel)]="queryString" type="text" aria-label="search" class="form-control mb-2" placeholder="Search by task number / last name / registration number / community ..." (keydown.enter)="search()">
-    </div>
-    <div class="col-auto">
-      <button class="btn btn-primary" (click)="search()">SEARCH</button>
-    </div>
-  </div>
-  <!-- TODO: <a class="small" href="">advanced search</a> -->
-  <!-- TODO: show icons for HOH, RESTRICTED FILE -->
-  <div class="clearfix my-3">
-    <div class="float-right">
-      <app-font-awesome-icon class="mr-1" icon="user"></app-font-awesome-icon>
-      <span class="ml-2">Head of Household</span>
-    </div>
-  </div>
-
-  <div class="space-between align-items-center my-3">
-    <section>
-      <app-pagination-summary [value]="resultsAndPagination?.metadata">
-        <!-- <span class="font-weight-bold">ACTIVE</span> evacuees -->
-        <span class="font-weight-bold">ACTIVE</span> registrations
-        <!-- TODO: <small class="ml-3">(<a href="">show deactivated</a>)</small> -->
-      </app-pagination-summary>
-    </section>
-    <section>
-      <select class="form-control" [(ngModel)]="sort" (change)="search()">
-        <!-- <option *ngFor="let increment of increments" [value]="increment">{{increment}}</option> -->
-        <option [value]="'-registrationCompletionDate'">Sort by: Latest (default)</option>
-        <option [value]="'headOfHousehold.lastName'">Last Name</option>
-        <option [value]="'hostCommunity.name'">Community Evacuated to</option>
-        <option [value]="'incidentTask.community.name'">Community Evacuated from</option>
-      </select>
-    </section>
-  </div>
-  <div class="search-results">
-    <div *ngIf="rows.length; then searchResultsTable; else notFound"></div>
-  </div>
-
-  <ng-template #searchResultsTable>
-    <pre>{{evacuees|json}}</pre>
-    <!-- <table class="table">
-      <thead class="thead-light">
-        <th scope="col"></th>
-        <th scope="col">Last Name</th>
-        <th scope="col">First Name</th>
-        <th scope="col">Task #</th>
-        <th scope="col">ESS File #</th>
-        <th scope="col">Evacuated From</th>
-        <th scope="col">Evacuated To</th>
-        <th scope="col">Action</th>
-      </thead>
-      <tbody>
-        <tr [class.table-danger]="rowItem.restrictedAccess" *ngFor="let rowItem of rows">
-          <th scope="row">
-            <app-font-awesome-icon class="mr-1 float-left" *ngIf="rowItem.headOfHousehold" icon="user">
-            </app-font-awesome-icon>
-            <app-font-awesome-icon class="float-right" *ngIf="rowItem.restrictedAccess" icon="lock">
-            </app-font-awesome-icon>
-          </th>
-          <td>{{rowItem.lastName}}</td>
-          <td>{{rowItem.firstName}}</td>
-          <td>{{rowItem.incidentTaskTaskNumber}}</td>
-          <td>{{rowItem.essFileNumber}}</td>
-          <td>{{rowItem.evacuatedFrom}}</td>
-          <td>{{rowItem.evacuatedTo}}</td>
-          <td>
-            <a href="javascript:void(0)" *ngIf="!rowItem.incidentTaskTaskNumber" (click)="finalize(rowItem)">Finalize</a>
-            <a href="javascript:void(0)" *ngIf="rowItem.incidentTaskTaskNumber" (click)="view(rowItem)">View</a>
-          </td>
-        </tr>
-      </tbody>
-    </table> -->
-  </ng-template>
-
-  <ng-template #notFound>
-    <p>{{notFoundMessage}}</p>
-  </ng-template>
-
-  <!-- <select class="form-control" [(ngModel)]="maxSize">
-    <option *ngFor="let increment of increments" [value]="increment">{{increment}}</option>
-  </select> -->
-  <div class="col clearfix" *ngIf="resultsAndPagination?.metadata?.totalPages > 1">
-    <div class="float-right">
-      <app-paginator [metaData]="resultsAndPagination?.metadata" (params)="onPaginationEvent($event)"></app-paginator>
-    </div>
-  </div>
-</div>
